==> 01_02_distance_calculation_others.ipynb <==
import matplotlib.pyplot as plt
import pandas as pd
import geopandas as gpd
import urbanpy as up
from tqdm.notebook import tqdm

tqdm.pandas()

gdf_celdas = gpd.read_parquet("outputs/celdas_countries_wo_br_02_06_2025.parquet")
print(f"{gdf_celdas.shape[0]} Cells loaded")

gdf_schools = gpd.read_parquet("outputs/amazon_schools.parquet")
print(f"{gdf_schools.shape[0]} Cells loaded")

gdf_schools.loc[:, "lat"] = gdf_schools["geometry"].y
gdf_schools.loc[:, "lon"] = gdf_schools["geometry"].x

gdf_celdas.head()

gdf_celdas["code"].value_counts()

gdf_celdas["code"].isna().sum()

gdf_celdas_peru = gdf_celdas[gdf_celdas["code"] == "per"]

gdf_schools["country_code"].value_counts()

gdf_schools_peru = gdf_schools[gdf_schools["country_code"] == "PER"]

assert (
    gdf_schools_peru.crs.to_string() == gdf_celdas.crs.to_string()
), "CRS do not match"

gdf_schools_peru.head()

gdf_primary_schools_peru = gdf_schools_peru[gdf_schools_peru["nivel_primaria"] == 1]
gdf_middle_schools_peru = gdf_schools_peru[gdf_schools_peru["nivel_media"] == 1]
gdf_secondary_schools_peru = gdf_schools_peru[gdf_schools_peru["nivel_secundaria"] == 1]

total_schools = gdf_schools_peru.shape[0]
print("total # of schools in peru", gdf_schools_peru.shape[0])
print("primary:", gdf_primary_schools_peru.shape[0])
print("middle:", gdf_middle_schools_peru.shape[0])
print("secondary:", gdf_secondary_schools_peru.shape[0])
print(
    "add primary, middle, and secondary:",
    gdf_primary_schools_peru.shape[0]
    + gdf_middle_schools_peru.shape[0]
    + gdf_secondary_schools_peru.shape[0],
)

# start server
up.routing.start_osrm_server("south-america/peru", "foot")

gdf_celdas_peru_access_primaria = up.accessibility.travel_times(
    gdf_celdas_peru, gdf_primary_schools_peru, "primary_schools"
)

gdf_celdas_peru_access_primaria

gdf_primary_schools_peru.reset_index()

# Obtain the school unique_id and lat_lon
gdf_celdas_peru_access_primaria = pd.merge(
    gdf_celdas_peru_access_primaria,
    gdf_primary_schools_peru.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_primary_schools_ix",
    right_index=True,
    suffixes=("", "_primary_school"),
)

gdf_celdas_peru_access_middle = up.accessibility.travel_times(
    gdf_celdas_peru_access_primaria, gdf_middle_schools_peru, "middle_schools"
)

# Obtain the middle school unique_id and lat_lon
gdf_celdas_peru_access_middle = pd.merge(
    gdf_celdas_peru_access_middle,
    gdf_middle_schools_peru.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_middle_schools_ix",
    right_index=True,
    suffixes=("", "_middle_school"),
)

gdf_celdas_peru_access_middle.head()

gdf_celdas_peru_access_total = up.accessibility.travel_times(
    gdf_celdas_peru_access_middle, gdf_secondary_schools_peru, "secondary_schools"
)

# Obtain the secondary school unique_id and lat_lon
gdf_celdas_peru_access_total = pd.merge(
    gdf_celdas_peru_access_total,
    gdf_secondary_schools_peru.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_secondary_schools_ix",
    right_index=True,
    suffixes=("", "_secondary_school"),
)

gdf_celdas_peru_access_total.head()

gdf_celdas_peru_access_total["category"].value_counts(), gdf_celdas_peru_access_total[
    "category"
].isna().sum()

import seaborn as sns

# Group by settlement type (smod) and calculate mean durations
grouped = (
    gdf_celdas_peru_access_total.groupby("category")[
        ["duration_to_nearest_primary_schools", "duration_to_nearest_secondary_schools"]
    ]
    .mean()
    .reset_index()
)

# Plot
grouped_melted = grouped.melt(
    id_vars="category", var_name="School Type", value_name="Avg Duration (min)"
)
sns.barplot(
    data=grouped_melted, x="category", y="Avg Duration (min)", hue="School Type"
)
plt.title("Average Travel Duration to Nearest Primary Schools by Settlement Type")
plt.ylabel("Minutes")
plt.xlabel("Settlement Type")
plt.xticks(rotation=45)
plt.tight_layout()
plt.show()

fig, axes = plt.subplots(1, 3, figsize=(15, 6), sharey=True)

# Plot for primary schools
gdf_celdas_peru_access_total[
    "duration_to_nearest_primary_schools_label"
].value_counts().plot(kind="bar", ax=axes[0])
axes[0].set_title("Distribution of Access Durations to Primary Schools")
axes[0].set_xlabel("Duration Category")
axes[0].set_ylabel("Number of Cells")
# axes[0].set_xticklabels(gdf_celdas_peru_access_total['duration_to_nearest_primary_schools_label'].unique().index, rotation=45)

# Plot for middle schools
gdf_celdas_peru_access_total[
    "duration_to_nearest_middle_schools_label"
].value_counts().plot(kind="bar", ax=axes[1], color="green")
axes[1].set_title("Distribution of Access Durations to Middle Schools")
axes[1].set_xlabel("Duration Category")

# Plot for secondary schools
gdf_celdas_peru_access_total[
    "duration_to_nearest_secondary_schools_label"
].value_counts().plot(kind="bar", ax=axes[2], color="orange")
axes[2].set_title("Distribution of Access Durations to Secondary Schools")
axes[2].set_xlabel("Duration Category")
# axes[1].set_xticklabels(secondary_labels.index, rotation=45)

axes[0].grid(True)
axes[1].grid(True)
axes[2].grid(True)

plt.tight_layout()
plt.show()

gdf_celdas_peru_access_total.head()

# Save the results
# gdf_celdas_peru_access.to_parquet("outputs/celdas_peru_access.parquet")
gdf_celdas_peru_access_total.to_parquet("outputs/celdas_peru_access_v3.parquet")

up.routing.stop_osrm_server("south-america/peru", "foot")

gdf_celdas["code"].value_counts()

gdf_schools["country_code"].value_counts()

# Select the data for the country
gdf_celdas_bol = gdf_celdas[gdf_celdas["code"] == "bol"]
gdf_schools_bol = gdf_schools[gdf_schools["country_code"] == "BOL"]

gdf_schools_bol

gdf_primary_schools_bol = gdf_schools_bol[gdf_schools_bol["nivel_primaria"] == 1]
gdf_middle_schools_bol = gdf_schools_bol[gdf_schools_bol["nivel_media"] == 1]
gdf_secondary_schools_bol = gdf_schools_bol[gdf_schools_bol["nivel_secundaria"] == 1]

print("total # of schools in bolivia", gdf_schools_bol.shape[0])
print("primary:", gdf_primary_schools_bol.shape[0])
print("middle:", gdf_middle_schools_bol.shape[0])
print("secondary:", gdf_secondary_schools_bol.shape[0])
print(
    "add primary, middle, and secondary:",
    gdf_primary_schools_bol.shape[0]
    + gdf_middle_schools_bol.shape[0]
    + gdf_secondary_schools_bol.shape[0],
)

up.routing.start_osrm_server("south-america/bolivia", "foot")

gdf_celdas_bol_access_primary = up.accessibility.travel_times(
    gdf_celdas_bol, gdf_primary_schools_bol, "primary_schools"
)

# Obtain the school unique_id and lat_lon
gdf_celdas_bol_access_primary = pd.merge(
    gdf_celdas_bol_access_primary,
    gdf_primary_schools_bol.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_primary_schools_ix",
    right_index=True,
    suffixes=("", "_primary_school"),
)

gdf_celdas_bol_access_middle = up.accessibility.travel_times(
    gdf_celdas_bol_access_primary, gdf_middle_schools_bol, "middle_schools"
)

# Obtain the middle school unique_id and lat_lon
gdf_celdas_bol_access_middle = pd.merge(
    gdf_celdas_bol_access_middle,
    gdf_middle_schools_bol.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_middle_schools_ix",
    right_index=True,
    suffixes=("", "_middle_school"),
)

gdf_celdas_bol_access_middle.head()

gdf_celdas_bol_access_total = up.accessibility.travel_times(
    gdf_celdas_bol_access_middle, gdf_secondary_schools_bol, "secondary_schools"
)

# Obtain the secondary school unique_id and lat_lon
gdf_celdas_bol_access_total = pd.merge(
    gdf_celdas_bol_access_total,
    gdf_secondary_schools_bol.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_secondary_schools_ix",
    right_index=True,
    suffixes=("", "_secondary_school"),
)

gdf_celdas_bol_access_total

up.routing.stop_osrm_server("south-america/bolivia", "foot")
# gdf_celdas_bol_access.head()

# Save the results
# gdf_celdas_bol_access.to_parquet("outputs/celdas_bol_access.parquet")
gdf_celdas_bol_access_total.to_parquet("outputs/celdas_bol_access_v3.parquet")

# Select the data for the country
gdf_celdas_ecu = gdf_celdas[gdf_celdas["code"] == "ecu"]
gdf_schools_ecu = gdf_schools[gdf_schools["country_code"] == "ECU"]

print("Ecuador data:")
print(f"{gdf_celdas_ecu.shape[0]} Cells loaded")
print(f"{gdf_schools_ecu.shape[0]} Schools loaded")

gdf_primary_schools_ecu = gdf_schools_ecu[gdf_schools_ecu["nivel_primaria"] == 1]
gdf_middle_schools_ecu = gdf_schools_ecu[gdf_schools_ecu["nivel_media"] == 1]
gdf_secondary_schools_ecu = gdf_schools_ecu[gdf_schools_ecu["nivel_secundaria"] == 1]

print("total # of schools in ecuador", gdf_schools_ecu.shape[0])
print("primary:", gdf_primary_schools_ecu.shape[0])
print("middle:", gdf_middle_schools_ecu.shape[0])
print("secondary:", gdf_secondary_schools_ecu.shape[0])
print(
    "add primary, middle, and secondary:",
    gdf_primary_schools_ecu.shape[0]
    + gdf_middle_schools_ecu.shape[0]
    + gdf_secondary_schools_ecu.shape[0],
)

# start server
up.routing.start_osrm_server("south-america/ecuador", "foot")

gdf_celdas_ecu_access_primary = up.accessibility.travel_times(
    gdf_celdas_ecu, gdf_primary_schools_ecu, "primary_schools"
)

gdf_celdas_ecu_access_primary.head()

print(
    "# of cells without access:",
    gdf_celdas_ecu_access_primary["duration_to_nearest_primary_schools_label"]
    .isna()
    .sum(),
)

print("# of cells by duration bin:")
print(
    gdf_celdas_ecu_access_primary[
        "duration_to_nearest_primary_schools_label"
    ].value_counts()
)

# Obtain the school unique_id and lat_lon
gdf_celdas_ecu_access_primary = pd.merge(
    gdf_celdas_ecu_access_primary,
    gdf_primary_schools_ecu.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_primary_schools_ix",
    right_index=True,
    suffixes=("", "_primary_school"),
)

gdf_celdas_ecu_access_middle = up.accessibility.travel_times(
    gdf_celdas_ecu_access_primary, gdf_middle_schools_ecu, "middle_schools"
)

gdf_celdas_ecu_access_middle.head()

print(
    "# of cells without access:",
    gdf_celdas_ecu_access_middle["duration_to_nearest_middle_schools_label"]
    .isna()
    .sum(),
)

print("# of cells by duration bin:")
print(
    gdf_celdas_ecu_access_middle[
        "duration_to_nearest_middle_schools_label"
    ].value_counts()
)

# Obtain the school unique_id and lat_lon
gdf_celdas_ecu_access_middle = pd.merge(
    gdf_celdas_ecu_access_middle,
    gdf_middle_schools_ecu.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_middle_schools_ix",
    right_index=True,
    suffixes=("", "_middle_school"),
)

gdf_celdas_ecu_access_total = up.accessibility.travel_times(
    gdf_celdas_ecu_access_middle, gdf_secondary_schools_ecu, "secondary_schools"
)

print(
    "# of cells without access:",
    gdf_celdas_ecu_access_total["duration_to_nearest_secondary_schools_label"]
    .isna()
    .sum(),
)

print("# of cells by duration bin:")
print(
    gdf_celdas_ecu_access_total[
        "duration_to_nearest_secondary_schools_label"
    ].value_counts()
)

# Obtain the secondary school unique_id and lat_lon
gdf_celdas_ecu_access_total = pd.merge(
    gdf_celdas_ecu_access_total,
    gdf_secondary_schools_ecu.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_secondary_schools_ix",
    right_index=True,
    suffixes=("", "_secondary_school"),
)

# gdf_celdas_ecu_access.to_parquet("outputs/celdas_ecu_access.parquet")
gdf_celdas_ecu_access_total.to_parquet("outputs/celdas_ecu_access_v3.parquet")

up.routing.stop_osrm_server("south-america/ecuador", "foot")

gdf_celdas_ecu_access_total.head()

# Select the data for the country
gdf_celdas_col = gdf_celdas[gdf_celdas["code"] == "col"]
gdf_schools_col = gdf_schools[gdf_schools["country_code"] == "COL"]

gdf_primary_schools_col = gdf_schools_col[gdf_schools_col["nivel_primaria"] == 1]
gdf_middle_schools_col = gdf_schools_col[gdf_schools_col["nivel_media"] == 1]
gdf_secondary_schools_col = gdf_schools_col[gdf_schools_col["nivel_secundaria"] == 1]

print("total # of schools in colombia", gdf_schools_col.shape[0])
print("primary:", gdf_primary_schools_col.shape[0])
print("middle:", gdf_middle_schools_col.shape[0])
print("secondary:", gdf_secondary_schools_col.shape[0])
print(
    "add primary, middle, and secondary:",
    gdf_primary_schools_col.shape[0]
    + gdf_middle_schools_col.shape[0]
    + gdf_secondary_schools_col.shape[0],
)

# start server
up.routing.start_osrm_server("south-america/colombia", "foot")

gdf_celdas_col_access_primary = up.accessibility.travel_times(
    gdf_celdas_col, gdf_primary_schools_col, "primary_schools"
)

print(
    "# of cells without access:",
    gdf_celdas_col_access_primary["duration_to_nearest_primary_schools_label"]
    .isna()
    .sum(),
)

print("# of cells by duration bin:")
print(
    gdf_celdas_col_access_primary[
        "duration_to_nearest_primary_schools_label"
    ].value_counts()
)

# Obtain the school unique_id and lat_lon
gdf_celdas_col_access_primary = pd.merge(
    gdf_celdas_col_access_primary,
    gdf_primary_schools_col.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_primary_schools_ix",
    right_index=True,
    suffixes=("", "_primary_school"),
)

gdf_celdas_col_access_middle = up.accessibility.travel_times(
    gdf_celdas_col_access_primary, gdf_middle_schools_col, "middle_schools"
)

print(
    "# of cells without access:",
    gdf_celdas_col_access_middle["duration_to_nearest_middle_schools_label"]
    .isna()
    .sum(),
)

print("# of cells by duration bin:")
print(
    gdf_celdas_col_access_middle[
        "duration_to_nearest_middle_schools_label"
    ].value_counts()
)

# Obtain the school unique_id and lat_lon
gdf_celdas_col_access_middle = pd.merge(
    gdf_celdas_col_access_middle,
    gdf_middle_schools_col.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_middle_schools_ix",
    right_index=True,
    suffixes=("", "_middle_school"),
)

gdf_celdas_col_access_total = up.accessibility.travel_times(
    gdf_celdas_col_access_middle, gdf_secondary_schools_col, "secondary_schools"
)

print(
    "# of cells without access:",
    gdf_celdas_col_access_total["duration_to_nearest_secondary_schools_label"]
    .isna()
    .sum(),
)

print("# of cells by duration bin:")
print(
    gdf_celdas_col_access_total[
        "duration_to_nearest_secondary_schools_label"
    ].value_counts()
)

# Obtain the secondary school unique_id and lat_lon
gdf_celdas_col_access_total = pd.merge(
    gdf_celdas_col_access_total,
    gdf_secondary_schools_col.reset_index()[["id_edificio", "lat", "lon"]],
    how="left",
    left_on="nearest_secondary_schools_ix",
    right_index=True,
    suffixes=("", "_secondary_school"),
)

# gdf_celdas_ecu_access.to_parquet("outputs/celdas_ecu_access.parquet")
gdf_celdas_col_access_total.to_parquet("outputs/celdas_col_access_v3.parquet")

up.routing.stop_osrm_server("south-america/colombia", "foot")

gdf_celdas_col_access_total.head()

# Read and concatenate them into a single GeoDataFrame
gdf_combined_comp = gpd.GeoDataFrame(
    pd.concat(
        [
            gdf_celdas_bol_access_total,
            gdf_celdas_col_access_total,
            gdf_celdas_ecu_access_total,
            gdf_celdas_peru_access_total,
        ],
        ignore_index=True,
    )
)

# Display the combined GeoDataFrame
gdf_combined_comp.head()

# gdf_combined_comp["nivel_educativo"] = gdf_combined_comp.apply(
#     lambda x: (
#         "Primaria"
#         if x["EduNivelPrimaria"] == 1
#         else "Secundaria" if x["EduNivelSecundariaTotal"] == 1 else "Ninguno"
#     ),
#     axis=1,
# )

# gdf_combined_comp["nivel_educativo"].value_counts()

# Save the combined GeoDataFrame to a new file
gdf_combined_comp.to_parquet("outputs/celdas_combined_access_v3.parquet")

gdf_combined_comp[
    [
        "duration_to_nearest_primary_schools_label",
        "duration_to_nearest_middle_schools_label",
        "duration_to_nearest_secondary_schools_label",
    ]
].describe().T

gdf_combined_comp[
    [
        "duration_to_nearest_primary_schools_label",
        "duration_to_nearest_middle_schools_label",
        "duration_to_nearest_secondary_schools_label",
    ]
].isna().sum()

gdf_combined_comp[
    [
        "duration_to_nearest_primary_schools_label",
        "duration_to_nearest_middle_schools_label",
        "duration_to_nearest_secondary_schools_label",
    ]
].isna().sum() / gdf_combined_comp.shape[0] * 100

gdf_combined_comp[
    [
        "duration_to_nearest_primary_schools",
        "duration_to_nearest_middle_schools",
        "duration_to_nearest_secondary_schools",
    ]
].isna().sum()

gdf_combined_comp[
    [
        "duration_to_nearest_primary_schools",
        "duration_to_nearest_middle_schools",
        "duration_to_nearest_secondary_schools",
    ]
].describe()

(
    gdf_combined_comp[
        [
            "duration_to_nearest_primary_schools",
            "duration_to_nearest_middle_schools",
            "duration_to_nearest_secondary_schools",
        ]
    ]
    == 0
).sum()

gdf_combined_comp[gdf_combined_comp["duration_to_nearest_primary_schools"] == 0][
    [
        "cell_id",
        "lat",
        "lon",
        "code",
        "lat_primary_school",
        "lon_primary_school",
        "distance_to_nearest_primary_schools",
        "duration_to_nearest_primary_schools",
        "nearest_primary_schools_ix",
    ]
]

gdf_combined_comp[
    gdf_combined_comp["duration_to_nearest_primary_schools"] == 0
].to_parquet("cells_with_0_distance.parquet")

# NOTE: duration and distance values equal to zero represent "No access"
