==> 05_01_new_schools.ipynb <==
import geopandas as gpd

amazon_schools = gpd.read_file("inputs/new_schools/escuelas_amazonia.geojson")

amazon_schools.crs

amazon_schools.head()

amazon_schools["country_code"] = amazon_schools["id_edificio"].str[:3]

amazon_schools["country_code"].value_counts()

amazon_schools.to_parquet(
    "outputs/amazon_schools.parquet",
    index=False,
)
